Modulise virtual-league page for re-use
